+++ packages/playground-ui/src/index.ts
@@ -3,6 +3,7 @@ import './index.css';
 export * from './contexts/mastra-client-context';
 export * from './domains/agents/index';
 export * from './domains/networks/index';
+export * from './domains/tools/index';
 export * from './domains/workflows/index';
 export * from './domains/resizable-panel';
 export * from './components/dynamic-form/index';
